+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Changed
 - Move IRA table columns [#338](https://github.com/policy-design-lab/pdl-frontend/issues/338)
 - Upgraded the node version from 16 to 22, including Dockerfile and GitHub Actions [#352](https://github.com/policy-design-lab/pdl-frontend/issues/352)
+- Body description of issue brief component to a parameter [#358](https://github.com/policy-design-lab/pdl-frontend/issues/358)
 
 ### Removed
 - Remove unused math functions [#71](https://github.com/policy-design-lab/pdl-frontend/issues/71)